Cloudinit can now return the cloudinit configuration as a bash script. SSHInit only executes it over ssh
