+++ apiserver/client/client_test.go
@@ -3377,20 +3377,20 @@ func (s *clientSuite) TestProvisioningScript(c *gc.C) {
 	c.Assert(err, jc.ErrorIsNil)
 	icfg, err := client.InstanceConfig(s.State, machineId, apiParams.Nonce, "")
 	c.Assert(err, jc.ErrorIsNil)
-	sshinitScript, err := manual.ProvisioningScript(icfg)
+	provisioningScript, err := manual.ProvisioningScript(icfg)
 	c.Assert(err, jc.ErrorIsNil)
 	// ProvisioningScript internally calls MachineConfig,
 	// which allocates a new, random password. Everything
 	// about the scripts should be the same other than
 	// the line containing "oldpassword" from agent.conf.
 	scriptLines := strings.Split(script, "\n")
-	sshinitScriptLines := strings.Split(sshinitScript, "\n")
-	c.Assert(scriptLines, gc.HasLen, len(sshinitScriptLines))
+	provisioningScriptLines := strings.Split(provisioningScript, "\n")
+	c.Assert(scriptLines, gc.HasLen, len(provisioningScriptLines))
 	for i, line := range scriptLines {
 		if strings.Contains(line, "oldpassword") {
 			continue
 		}
-		c.Assert(line, gc.Equals, sshinitScriptLines[i])
+		c.Assert(line, gc.Equals, provisioningScriptLines[i])
 	}
 }
