+++ cloudconfig/cloudinit/packaging.go
@@ -5,10 +5,7 @@
 package cloudinit
 
 import (
-	"fmt"
-
-	"github.com/juju/juju/cloudconfig/cloudinit/packaging"
-	"github.com/juju/utils/apt"
+	"github.com/juju/juju/version"
 	"github.com/juju/utils/proxy"
 )
 
@@ -31,17 +28,20 @@ func AddPackageCommands(
 	series string,
 	aptProxySettings proxy.Settings,
 	aptMirror string,
-	c CloudConfig,
+	cfg CloudConfig,
 	addUpdateScripts bool,
 	addUpgradeScripts bool,
-) {
+) error {
 	// Check preconditions
-	if c == nil {
-		panic("c is nil")
+	if cfg == nil {
+		panic("AddPackageCommands received nil CloudConfig")
 	}
 
 	// Set the APT mirror.
-	c.SetPackageMirror(aptMirror)
+	// TODO: in the future we might pass yumMirror as well here
+	// SetPackage mirror knows the OS of the configuration we need to make sure
+	// what we pass to it or parse it inside
+	cfg.SetPackageMirror(aptMirror)
 
 	// For LTS series which need support for the cloud-tools archive,
 	// we need to enable apt-get update regardless of the environ
@@ -51,135 +51,70 @@ func AddPackageCommands(
 	}
 
 	// Bring packages up-to-date.
-	c.SetSystemUpdate(addUpdateScripts)
-	c.SetSystemUpgrade(addUpgradeScripts)
+	cfg.SetSystemUpdate(addUpdateScripts)
+	cfg.SetSystemUpgrade(addUpgradeScripts)
 	//c.SetAptGetWrapper("eatmydata")
 
 	// If we're not doing an update, adding these packages is
 	// meaningless.
+	// TODO: Decide when we update on CentOS
 	if addUpdateScripts {
-		requiredPackages := []string{
-			"curl",
-			"cpu-checker",
-			// TODO(axw) 2014-07-02 #1277359
-			// Don't install bridge-utils in cloud-init;
-			// leave it to the networker worker.
-			"bridge-utils",
-			"rsyslog-gnutls",
-			"cloud-utils",
-			"cloud-image-utils",
-		}
-
-		// The required packages need to come from the correct repo.
-		// For precise, that might require an explicit --target-release parameter.
-		for _, pkg := range requiredPackages {
-			// We cannot just pass requiredPackages below, because
-			// this will generate install commands which older
-			// versions of cloud-init (e.g. 0.6.3 in precise) will
-			// interpret incorrectly (see bug http://pad.lv/1424777).
-			cmds := apt.GetPreparePackages([]string{pkg}, series)
-			if len(cmds) != 1 {
-				// One package given, one command (with possibly
-				// multiple args) expected.
-				panic(fmt.Sprintf("expected one install command per package, got %v", cmds))
-			}
-			for _, p := range cmds[0] {
-				// We need to add them one package at a time. Also for
-				// precise where --target-release
-				// precise-updates/cloud-tools is needed for some packages
-				// we need to pass these as "packages", otherwise the
-				// aforementioned older cloud-init will also fail to
-				// interpret the command correctly.
-				c.AddPackage(p)
-			}
+		err := updatePackages(series, cfg)
+		if err != nil {
+			return err
 		}
 	}
 
-	// Write out the apt proxy settings
-	if (aptProxySettings != proxy.Settings{}) {
-		filename := apt.ConfFile
-		c.AddBootCmd(fmt.Sprintf(
-			`printf '%%s\n' %s > %s`,
-			shquote(apt.ProxyContent(aptProxySettings)),
-			filename))
+	// TODO: Deal with proxy settings on CentOS
+	err := updateProxySettings(series, cfg, aptProxySettings)
+	if err != nil {
+		return err
 	}
+	return nil
 }
 
-// CloudToolsPrefsPath defines the default location of
-// apt_preferences(5) file for the cloud-tools pocket.
-const CloudToolsPrefsPath = "/etc/apt/preferences.d/50-cloud-tools"
+//TODO: These 3 functions might have to be refactored into something else, now
+//they're just ugly
+func updatePackages(series string, cfg CloudConfig) error {
+	os, err := version.GetOSFromSeries(series)
+	if err != nil {
+		return err
+	}
+	switch os {
+	case version.Ubuntu:
+		updatePackagesUbuntu(cfg, series)
+	case version.CentOS:
+		//TODO: Do something here
+	}
+	return nil
+}
+
+func updateProxySettings(series string, cfg CloudConfig, aptProxySettings proxy.Settings) error {
+	os, err := version.GetOSFromSeries(series)
+	if err != nil {
+		return err
+	}
+	switch os {
+	case version.Ubuntu:
+		updateProxySettingsUbuntu(cfg, aptProxySettings)
+	case version.CentOS:
+		//TODO: Do something here
+	}
+	return nil
+}
 
 // MaybeAddCloudArchiveCloudTools adds the cloud-archive cloud-tools
 // pocket to apt sources, if the series requires it.
-func MaybeAddCloudArchiveCloudTools(c CloudConfig, series string) {
-	if series != "precise" {
-		// Currently only precise; presumably we'll
-		// need to add each LTS in here as they're
-		// added to the cloud archive.
-		return
+func MaybeAddCloudArchiveCloudTools(cfg CloudConfig, series string) error {
+	os, err := version.GetOSFromSeries(series)
+	if err != nil {
+		return err
 	}
-	const url = "http://ubuntu-cloud.archive.canonical.com/ubuntu"
-	name := fmt.Sprintf("deb %s %s-updates/cloud-tools main", url, series)
-	prefs := packaging.PackagePreferences{
-		Path:        CloudToolsPrefsPath,
-		Explanation: "Pin with lower priority, not to interfere with charms",
-		Package:     "*",
-		Pin:         fmt.Sprintf("release n=%s-updates/cloud-tools", series),
-		Priority:    400,
+	switch os {
+	case version.Ubuntu:
+		maybeAddCloudArchiveCloudToolsUbuntu(cfg, series)
+	case version.CentOS:
+		//TODO: Do something here
 	}
-	c.AddPackagePreferences(prefs)
-	source := packaging.Source{
-		Url: name,
-		Key: CanonicalCloudArchiveSigningKey,
-	}
-	c.AddPackageSource(source)
+	return nil
 }
-
-const CanonicalCloudArchiveSigningKey = `-----BEGIN PGP PUBLIC KEY BLOCK-----
-Version: SKS 1.1.4
-Comment: Hostname: keyserver.ubuntu.com
-
-mQINBFAqSlgBEADPKwXUwqbgoDYgR20zFypxSZlSbrttOKVPEMb0HSUx9Wj8VvNCr+mT4E9w
-Ayq7NTIs5ad2cUhXoyenrjcfGqK6k9R6yRHDbvAxCSWTnJjw7mzsajDNocXC6THKVW8BSjrh
-0aOBLpht6d5QCO2vyWxw65FKM65GOsbX03ZngUPMuOuiOEHQZo97VSH2pSB+L+B3d9B0nw3Q
-nU8qZMne+nVWYLYRXhCIxSv1/h39SXzHRgJoRUFHvL2aiiVrn88NjqfDW15HFhVJcGOFuACZ
-nRA0/EqTq0qNo3GziQO4mxuZi3bTVL5sGABiYW9uIlokPqcS7Fa0FRVIU9R+bBdHZompcYnK
-AeGag+uRvuTqC3MMRcLUS9Oi/P9I8fPARXUPwzYN3fagCGB8ffYVqMunnFs0L6td08BgvWwe
-r+Buu4fPGsQ5OzMclgZ0TJmXyOlIW49lc1UXnORp4sm7HS6okA7P6URbqyGbaplSsNUVTgVb
-i+vc8/jYdfExt/3HxVqgrPlq9htqYgwhYvGIbBAxmeFQD8Ak/ShSiWb1FdQ+f7Lty+4mZLfN
-8x4zPZ//7fD5d/PETPh9P0msF+lLFlP564+1j75wx+skFO4v1gGlBcDaeipkFzeozndAgpeg
-ydKSNTF4QK9iTYobTIwsYfGuS8rV21zE2saLM0CE3T90aHYB/wARAQABtD1DYW5vbmljYWwg
-Q2xvdWQgQXJjaGl2ZSBTaWduaW5nIEtleSA8ZnRwbWFzdGVyQGNhbm9uaWNhbC5jb20+iQI3
-BBMBCAAhBQJQKkpYAhsDBQsJCAcDBRUKCQgLBRYCAwEAAh4BAheAAAoJEF7bG2LsSSbqKxkQ
-AIKtgImrk02YCDldg6tLt3b69ZK0kIVI3Xso/zCBZbrYFmgGQEFHAa58mIgpv5GcgHHxWjpX
-3n4tu2RM9EneKvFjFBstTTgoyuCgFr7iblvs/aMW4jFJAiIbmjjXWVc0CVB/JlLqzBJ/MlHd
-R9OWmojN9ZzoIA+i+tWlypgUot8iIxkR6JENxit5v9dN8i6anmnWybQ6PXFMuNi6GzQ0JgZI
-Vs37n0ks2wh0N8hBjAKuUgqu4MPMwvNtz8FxEzyKwLNSMnjLAhzml/oje/Nj1GBB8roj5dmw
-7PSul5pAqQ5KTaXzl6gJN5vMEZzO4tEoGtRpA0/GTSXIlcx/SGkUK5+lqdQIMdySn8bImU6V
-6rDSoOaI9YWHZtpv5WeUsNTdf68jZsFCRD+2+NEmIqBVm11yhmUoasC6dYw5l9P/PBdwmFm6
-NBUSEwxb+ROfpL1ICaZk9Jy++6akxhY//+cYEPLin02r43Z3o5Piqujrs1R2Hs7kX84gL5Sl
-BzTM4Ed+ob7KVtQHTefpbO35bQllkPNqfBsC8AIC8xvTP2S8FicYOPATEuiRWs7Kn31TWC2i
-wswRKEKVRmN0fdpu/UPdMikyoNu9szBZRxvkRAezh3WheJ6MW6Fmg9d+uTFJohZt5qHdpxYa
-4beuN4me8LF0TYzgfEbFT6b9D6IyTFoT0LequQINBFAqSlgBEADmL3TEq5ejBYrA+64zo8FY
-vCF4gziPa5rCIJGZ/gZXQ7pm5zek/lOe9C80mhxNWeLmrWMkMOWKCeaDMFpMBOQhZZmRdakO
-nH/xxO5x+fRdOOhy+5GTRJiwkuGOV6rB9eYJ3UN9caP2hfipCMpJjlg3j/GwktjhuqcBHXhA
-HMhzxEOIDE5hmpDqZ051f8LGXld9aSL8RctoYFM8sgafPVmICTCq0Wh03dr5c2JAgEXy3ush
-Ym/8i2WFmyldo7vbtTfx3DpmJc/EMpGKV+GxcI3/ERqSkde0kWlmfPZbo/5+hRqSryqfQtRK
-nFEQgAqAhPIwXwOkjCpPnDNfrkvzVEtl2/BWP/1/SOqzXjk9TIb1Q7MHANeFMrTCprzPLX6I
-dC4zLp+LpV91W2zygQJzPgWqH/Z/WFH4gXcBBqmI8bFpMPONYc9/67AWUABo2VOCojgtQmjx
-uFn+uGNw9PvxJAF3yjl781PVLUw3n66dwHRmYj4hqxNDLywhhnL/CC7KUDtBnUU/CKn/0Xgm
-9oz3thuxG6i3F3pQgpp7MeMntKhLFWRXo9Bie8z/c0NV4K5HcpbGa8QPqoDseB5WaO4yGIBO
-t+nizM4DLrI+v07yXe3Jm7zBSpYSrGarZGK68qamS3XPzMshPdoXXz33bkQrTPpivGYQVRZu
-zd/R6b+6IurV+QARAQABiQIfBBgBCAAJBQJQKkpYAhsMAAoJEF7bG2LsSSbq59EP/1U3815/
-yHV3cf/JeHgh6WS/Oy2kRHp/kJt3ev/l/qIxfMIpyM3u/D6siORPTUXHPm3AaZrbw0EDWByA
-3jHQEzlLIbsDGZgrnl+mxFuHwC1yEuW3xrzgjtGZCJureZ/BD6xfRuRcmvnetAZv/z98VN/o
-j3rvYhUi71NApqSvMExpNBGrdO6gQlI5azhOu8xGNy4OSke8J6pAsMUXIcEwjVEIvewJuqBW
-/3rj3Hh14tmWjQ7shNnYBuSJwbLeUW2e8bURnfXETxrCmXzDmQldD5GQWCcD5WDosk/HVHBm
-Hlqrqy0VO2nE3c73dQlNcI4jVWeC4b4QSpYVsFz/6Iqy5ZQkCOpQ57MCf0B6P5nF92c5f3TY
-PMxHf0x3DrjDbUVZytxDiZZaXsbZzsejbbc1bSNp4hb+IWhmWoFnq/hNHXzKPHBTapObnQju
-+9zUlQngV0BlPT62hOHOw3Pv7suOuzzfuOO7qpz0uAy8cFKe7kBtLSFVjBwaG5JX89mgttYW
-+lw9Rmsbp9Iw4KKFHIBLOwk7s+u0LUhP3d8neBI6NfkOYKZZCm3CuvkiOeQP9/2okFjtj+29
-jEL+9KQwrGNFEVNe85Un5MJfYIjgyqX3nJcwypYxidntnhMhr2VD3HL2R/4CiswBOa4g9309
-p/+af/HU1smBrOfIeRoxb8jQoHu3
-=xg4S
------END PGP PUBLIC KEY BLOCK-----`
