+++ cloudconfig/containerinit/export_test.go
@@ -1,0 +1,12 @@
+// Copyright 2015 Canonical Ltd.
+// Copyright 2015 Cloudbase Solutions SRL
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package containerinit
+
+var (
+	CloudInitUserData              = cloudInitUserData
+	NetworkInterfacesFile          = &networkInterfacesFile
+	NewCloudInitConfigWithNetworks = newCloudInitConfigWithNetworks
+	ShutdownInitScript             = shutdownInitScript
+)
