--- cloudconfig/sshinit/export_test.go
@@ -1,6 +1,0 @@
-// Copyright 2014 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package sshinit
-
-const Aptget = aptget
