--- cloudconfig/sshinit/suite_test.go
@@ -1,14 +1,0 @@
-// Copyright 2013 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package sshinit_test
-
-import (
-	stdtesting "testing"
-
-	gc "gopkg.in/check.v1"
-)
-
-func Test(t *stdtesting.T) {
-	gc.TestingT(t)
-}
