+++ container/export_test.go
@@ -3,11 +3,6 @@
 
 package container
 
-var (
-	NetworkInterfacesFile = &networkInterfacesFile
-	CloudInitUserData     = cloudInitUserData
-)
-
 // IsLocked is used just to see if the local lock instance is locked, and
 // is only required for use in tests.
 func IsLocked(lock *Lock) bool {
