+++ provider/azure/customdata.go
@@ -7,14 +7,14 @@ import (
 	"encoding/base64"
 	"fmt"
 
-	"github.com/juju/juju/cloudconfig"
 	"github.com/juju/juju/cloudconfig/instancecfg"
+	"github.com/juju/juju/cloudconfig/providerinit"
 )
 
 // makeCustomData produces custom data for Azure.  This is a base64-encoded
 // zipfile of cloudinit userdata.
 func makeCustomData(cfg *instancecfg.InstanceConfig) (string, error) {
-	zipData, err := cloudconfig.ComposeUserData(cfg, nil)
+	zipData, err := providerinit.ComposeUserData(cfg, nil)
 	if err != nil {
 		return "", fmt.Errorf("failure while generating custom data: %v", err)
 	}
